+++ docs/changes.rst
@@ -5,6 +5,8 @@ Change Log
 * 0.2.7 (dev)
 
   * added :meth:`.walktree`, recursively walk, depth first, a remote directory structure.  Used as the base of :meth:`get_recursive`. See tests/test_walktree.py for examples.
+  * added :meth:`.unlink` as synonym for :meth:`.remove`
+
 
 * 0.2.6 (current, released 2014-05-17)